ironic-lib 2.0.0 release

meta:version: 2.0.0
meta:series: newton
meta:release-type: release
meta:announce: openstack-dev@lists.openstack.org
meta:pypi: yes
meta:first: no
meta:release:Author: Dmitry Tantsur <divius.inside@gmail.com>
meta:release:Commit: Dmitry Tantsur <divius.inside@gmail.com>
meta:release:Change-Id: I21707e84459d904247f5e7808fc9647fb1fc168a
meta:release:Code-Review+1: Jim Rollenhagen <jim@jimrollenhagen.com>
meta:release:Code-Review+2: Doug Hellmann <doug@doughellmann.com>
meta:release:Workflow+1: Doug Hellmann <doug@doughellmann.com>

+++ ironic_lib/disk_utils.py
@@ -45,17 +45,18 @@ opts = [
     cfg.IntOpt('efi_system_partition_size',
                default=200,
                help='Size of EFI system partition in MiB when configuring '
-                    'UEFI systems for local boot.',
-               deprecated_group='deploy'),
+                    'UEFI systems for local boot.'),
+    cfg.IntOpt('bios_boot_partition_size',
+               default=1,
+               help='Size of BIOS Boot partition in MiB when configuring '
+                    'GPT partitioned systems for local boot in BIOS.'),
     cfg.StrOpt('dd_block_size',
                default='1M',
-               help='Block size to use when writing to the nodes disk.',
-               deprecated_group='deploy'),
+               help='Block size to use when writing to the nodes disk.'),
     cfg.IntOpt('iscsi_verify_attempts',
                default=3,
                help='Maximum attempts to verify an iSCSI connection is '
-                    'active, sleeping 1 second between attempts.',
-               deprecated_group='deploy'),
+                    'active, sleeping 1 second between attempts.'),
 ]
 
 CONF = cfg.CONF
@@ -180,8 +181,13 @@ def make_partitions(dev, root_mb, swap_mb, ephemeral_mb,
     if boot_mode == "uefi" and boot_option == "local":
         part_num = dp.add_partition(CONF.disk_utils.efi_system_partition_size,
                                     fs_type='fat32',
-                                    bootable=True)
+                                    boot_flag='boot')
         part_dict['efi system partition'] = part_template % part_num
+
+    if boot_mode == "bios" and boot_option == "local" and disk_label == "gpt":
+        part_num = dp.add_partition(CONF.disk_utils.bios_boot_partition_size,
+                                    boot_flag='bios_grub')
+        part_dict['BIOS Boot partition'] = part_template % part_num
 
     if ephemeral_mb:
         LOG.debug("Add ephemeral partition (%(size)d MB) to device: %(dev)s "
@@ -208,8 +214,14 @@ def make_partitions(dev, root_mb, swap_mb, ephemeral_mb,
     LOG.debug("Add root partition (%(size)d MB) to device: %(dev)s "
               "for node %(node)s",
               {'dev': dev, 'size': root_mb, 'node': node_uuid})
-    part_num = dp.add_partition(root_mb, bootable=(boot_option == "local" and
-                                                   boot_mode == "bios"))
+
+    boot_val = None
+    if (boot_mode == "bios" and boot_option == "local" and
+        disk_label == "msdos"):
+        boot_val = 'boot'
+
+    part_num = dp.add_partition(root_mb, boot_flag=boot_val)
+
     part_dict['root'] = part_template % part_num
 
     if commit:
@@ -316,10 +328,19 @@ def destroy_disk_metadata(dev, node_uuid):
     # https://bugs.launchpad.net/ironic/+bug/1317647
     LOG.debug("Start destroy disk metadata for node %(node)s.",
               {'node': node_uuid})
-    utils.execute('wipefs', '--all', dev,
-                  run_as_root=True,
-                  check_exit_code=[0],
-                  use_standard_locale=True)
+    try:
+        utils.execute('wipefs', '--force', '--all', dev,
+                      run_as_root=True,
+                      use_standard_locale=True)
+    except processutils.ProcessExecutionError as e:
+        # NOTE(zhenguo): Check if --force option is supported for wipefs,
+        # if not, we should try without it.
+        if '--force' in str(e):
+            utils.execute('wipefs', '--all', dev,
+                          run_as_root=True,
+                          use_standard_locale=True)
+        else:
+            raise e
     LOG.info(_LI("Disk metadata on %(dev)s successfully destroyed for node "
                  "%(node)s"), {'dev': dev, 'node': node_uuid})
